+++ src/marais/graphql/dsl/SchemaBuilderContext.kt
@@ -2,6 +2,7 @@ package marais.graphql.dsl
 
 import org.apache.logging.log4j.Logger
 import kotlin.reflect.KClass
+import kotlin.reflect.full.isSubclassOf
 
 /**
  * Called when converting a string input to your Id class
@@ -11,12 +12,18 @@ typealias IdCoercer<T> = (value: String?) -> T?
 /**
  * Shared read-only context for all builders
  */
-abstract class SchemaBuilderContext(internal val log: Logger) {
-
-    abstract val idCoercers: Map<KClass<*>, IdCoercer<*>>
-    abstract val inputs: List<InputSpec>
+class SchemaBuilderContext(
+    val log: Logger,
+    val idCoercers: Map<KClass<*>, IdCoercer<*>>,
+    val inputs: List<InputSpec>,
+    val interfaces: List<InterfaceSpec<*>>
+) {
 
     internal fun getInputType(kclass: KClass<*>) = inputs.find { it.kclass == kclass }
 
     internal fun isInputType(kclass: KClass<*>) = getInputType(kclass) != null
+
+    internal fun getImplementedInterfaces(kclass: KClass<*>): List<KClass<*>> = interfaces.filter {
+        kclass.isSubclassOf(it.kclass)
+    }.map { it.kclass }
 }